public interface for State, BasicState implementation

+++ src/main/java/ru/nyakto/linguist/BasicStateFactory.java
@@ -1,14 +1,15 @@
 package ru.nyakto.linguist;
 
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.BiFunction;
 
 public class BasicStateFactory<T extends State> implements StateFactory<T> {
     private final FSM fsm;
-    private final BiFunction<FSM, Long, T> constructor;
-    private final AtomicLong idGenerator = new AtomicLong(0l);
+    private final BiFunction<FSM, Integer, T> constructor;
+    private final AtomicInteger idGenerator = new AtomicInteger(0);
 
-    public BasicStateFactory(FSM fsm, BiFunction<FSM, Long, T> constructor) {
+    public BasicStateFactory(FSM fsm, BiFunction<FSM, Integer, T> constructor) {
         this.fsm = fsm;
         this.constructor = constructor;
     }
