+++ src/main/java/ru/nyakto/linguist/FSM.java
@@ -7,18 +7,18 @@ import java.util.function.BiFunction;
 
 abstract public class FSM<T extends State, Symbol> implements Cloneable {
     private final T initialState;
-    private final Set<Long> finalStateIds = new HashSet<>();
-    protected final BiFunction<FSM, Long, T> stateConstructor;
+    private final Set<Integer> finalStateIds = new HashSet<>();
+    protected final BiFunction<FSM, Integer, T> stateConstructor;
     protected final StateFactory<T> stateFactory;
     protected final Set<T> states = new HashSet<>();
 
-    public FSM(BiFunction<FSM, Long, T> stateConstructor) {
+    public FSM(BiFunction<FSM, Integer, T> stateConstructor) {
         this.stateConstructor = stateConstructor;
         stateFactory = new BasicStateFactory<>(this, stateConstructor);
         initialState = createState();
     }
 
-    public BiFunction<FSM, Long, T> getStateConstructor() {
+    public BiFunction<FSM, Integer, T> getStateConstructor() {
         return stateConstructor;
     }
